Moving Mock to test: part 1
https://github.com/fog/fog-openstack/issues/222
