+++ lib/fog/identity/openstack.rb
@@ -32,15 +32,6 @@ module Fog
         service
       end
 
-      class Mock
-        attr_reader :config
-
-        def initialize(options = {})
-          @openstack_auth_uri = URI.parse(options[:openstack_auth_url])
-          @config = options
-        end
-      end
-
       class Real
         include Fog::OpenStack::Core
